requirements.txt
# OCR Options (choose one):
# Option 1: PaddleOCR (best for Chinese, but may have import issues on some systems)
# paddlepaddle>=2.5.0
# paddleocr>=2.7.0

# Option 2: EasyOCR (recommended alternative - more stable)
easyocr>=1.7.0

# Image processing
# Pillow>=10.0.0
# opencv-python>=4.8.0

# Data handling
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0
numpy>=1.24.0

# Feature extraction
rapidfuzz>=3.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0

# Image Embeddings (Session 2)
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0

# Model serving (Session 3)
joblib>=1.3.0
streamlit>=1.30.0
anthropic>=0.40.0
python-dotenv>=1.0.0
